refactor: reorder vue files
