+++ docs/src/components/search-results/AppSearchResults.vue
